Rewrite S3/WebDAV datastore to share code
The bulk of the logic is in ButlerURI so there is no
reason for the datastores dealing with remote resources
to be special themselves.

Currently backoff in S3 means that we have simple wrappers
there. Ideally backoff should move to ButlerURI
